\name{histmdl-package}
\alias{histmdl-package}
\alias{histmdl}
\docType{package}
\title{
  Informative Histograms-like Models
}
\description{
  This package adds support for parameter estimation in histogram-like
  models based on the MDL principle.
}
\details{
  \tabular{ll}{
    Package: \tab histmdl\cr
    Type:    \tab Package\cr
    Version: \tab 0.2\cr
    Date:    \tab 2013-01-30\cr
    License: \tab GPL\cr
  }
  See: \code{\link{hist.mdl}}.
}
\author{
Jouke Witteveen and Richard Gill

Maintainer: Jouke Witteveen <j.witteveen@gmail.com>
}
\references{
  TBA
}
\keyword{ package }
\seealso{
  \code{\link{hist.mdl}}
}
