\name{histmdl-package}
\alias{histmdl-package}
\docType{package}
\title{
  Informative Histograms-like Models
}
\description{
  This package adds support for parameter estimation in histogram-like
  models based on the MDL principle.
}
\details{
  \tabular{ll}{
    Package: \tab histmdl\cr
    Type:    \tab Package\cr
    Version: \tab 0.4\cr
    Date:    \tab 2015-03-16\cr
    License: \tab GPL\cr
  }
  See: \code{\link{histmdl}}.
}
\author{
Jouke Witteveen and Richard Gill

Maintainer: Jouke Witteveen <j.witteveen@gmail.com>
}
%\references{ TBA }
\keyword{ package }
\seealso{
  \code{\link{histmdl}}
}
